chore: refactor so rados changes more or less there but my functions still work
