+++ frontend/automation/e2e/wallet-edit.test.ts
@@ -2,10 +2,10 @@ import type { Page } from '@playwright/test'
 import { expect, test } from '@playwright/test'
 
 import data from '../data/test-data.json'
+import wallets from '../pages/wallets'
 import cleanup from '../support/cleanup'
 import { waitForNetworkConnected } from '../support/helpers'
 import initApp from '../support/init-app'
-import wallets from '../support/pages/wallets'
 import { restoreWallet } from '../support/wallet-api'
 
 const passphrase = data.testWalletPassphrase
