+++ frontend/automation/e2e/wallets.test.ts
@@ -2,9 +2,7 @@ import type { Page } from '@playwright/test'
 import { expect, test } from '@playwright/test'
 
 import data from '../data/test-data.json'
-import createWallet from '../pages/create-wallet'
-import viewWallet from '../pages/view-wallet'
-import wallets from '../pages/wallets'
+import { Pages } from '../pages/pages'
 import cleanup from '../support/cleanup'
 import {
   authenticate,
@@ -12,7 +10,6 @@ import {
   waitForNetworkConnected
 } from '../support/helpers'
 import initApp from '../support/init-app'
-import { Pages } from '../pages/pages'
 import { restoreWallet } from '../support/wallet-api'
 
 const passphrase = data.testWalletPassphrase
