+++ frontend/automation/playwright.config.ts
@@ -1,4 +1,5 @@
 import { defineConfig, devices } from '@playwright/test'
+import { useCallback } from 'react'
 /**
  * Read environment variables from file.
  * https://github.com/motdotla/dotenv
@@ -94,6 +95,7 @@ export default defineConfig({
   webServer: {
     command: process.env.CI ? 'yarn dev:test:ci' : 'yarn dev:test',
     port: 34115,
-    timeout: 6 * 60 * 1000
+    timeout: 6 * 60 * 1000,
+    reuseExistingServer: true
   }
 })
